requirements.txt
APScheduler==3.6.3
boto3==1.17.36
botocore==1.20.36
certifi==2020.12.5
jmespath==0.10.0
python-dateutil==2.8.1
python-telegram-bot==13.4.1
pytz==2021.1
s3transfer==0.3.6
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
